Checkpoint (wasm-bindgen success)

+++ chia-protocol/src/fullblock.rs
@@ -43,7 +43,7 @@ impl FullBlock {
         self.foliage.foliage_transaction_block_hash.is_some()
     }
 
-    pub fn total_iters(&self) -> u128 {
+    pub fn total_iters(&self) -> u64 {
         self.reward_chain_block.total_iters
     }
 
@@ -51,7 +51,7 @@ impl FullBlock {
         self.reward_chain_block.height
     }
 
-    pub fn weight(&self) -> u128 {
+    pub fn weight(&self) -> u64 {
         self.reward_chain_block.weight
     }
 
@@ -112,7 +112,7 @@ impl FullBlock {
 
     #[getter]
     #[pyo3(name = "total_iters")]
-    fn py_total_iters(&self) -> u128 {
+    fn py_total_iters(&self) -> u64 {
         self.total_iters()
     }
 
@@ -124,7 +124,7 @@ impl FullBlock {
 
     #[getter]
     #[pyo3(name = "weight")]
-    fn py_weight(&self) -> u128 {
+    fn py_weight(&self) -> u64 {
         self.weight()
     }
